Completed multiBlur interact, Fixed few compile warnings and errors on windows

+++ include/ofx/imageeffect.h
@@ -475,7 +475,7 @@ namespace ofx {
       
       struct BeginSequenceArgs : public SequenceArgs {
         FrameRange range;
-        int step;
+        double step;
         // supposed to be the same in the specs for EndSequence action
         
         BeginSequenceArgs(PropertySet &props);
